+++ pex/pex_bootstrapper.py
@@ -41,6 +41,7 @@ def iter_compatible_interpreters(
     path=None,  # type: Optional[str]
     valid_basenames=None,  # type: Optional[Iterable[str]]
     interpreter_constraints=None,  # type: Optional[Iterable[str]]
+    preferred_interpreter=None,  # type: Optional[PythonInterpreter]
 ):
     # type: (...) -> Iterator[PythonInterpreter]
     """Find all compatible interpreters on the system within the supplied constraints.
@@ -51,6 +52,8 @@ def iter_compatible_interpreters(
                             pypy, etc.).
     :param interpreter_constraints: Interpreter type and version constraint strings as described in
                                     `--interpreter-constraint`.
+    :param preferred_interpreter: For testing - an interpreter to prefer amongst all others.
+                                  Defaults to the current running interpreter.
 
     Interpreters are searched for in `path` if specified and $PATH if not.
 
@@ -72,30 +75,32 @@ def iter_compatible_interpreters(
         # type: () -> Iterator[InterpreterOrError]
         seen = set()
 
-        paths = None  # type: Optional[MutableSet[str]]
-        if path:
-            paths = OrderedSet(os.path.realpath(p) for p in path.split(os.pathsep))
+        normalized_paths = (
+            OrderedSet(os.path.realpath(p) for p in path.split(os.pathsep)) if path else None
+        )
 
-        current_interpreter = PythonInterpreter.get()
+        # Prefer the current interpreter, if valid.
+        current_interpreter = preferred_interpreter or PythonInterpreter.get()
         if not _valid_path or _valid_path(current_interpreter.binary):
-            if paths:
-                # Prefer the current interpreter if present on the `path`.
+            if normalized_paths:
                 candidate_paths = frozenset(
                     (current_interpreter.binary, os.path.dirname(current_interpreter.binary))
                 )
-                candidate_paths_in_path = candidate_paths.intersection(paths)
+                candidate_paths_in_path = candidate_paths.intersection(normalized_paths)
                 if candidate_paths_in_path:
-                    for p in candidate_paths_in_path:
-                        paths.remove(p)
+                    # In case the full path of the current interpreter binary was in the
+                    # `normalized_paths` we're searching, remove it to prevent identifying it again
+                    # just to then skip it as `seen`.
+                    normalized_paths.discard(current_interpreter.binary)
                     seen.add(current_interpreter)
                     yield current_interpreter
             else:
-                # We may have been invoked with a specific interpreter, make sure our sys.executable is
-                # included as a candidate in this case.
                 seen.add(current_interpreter)
                 yield current_interpreter
 
-        for interp in PythonInterpreter.iter_candidates(paths=paths, path_filter=_valid_path):
+        for interp in PythonInterpreter.iter_candidates(
+            paths=normalized_paths, path_filter=_valid_path
+        ):
             if interp not in seen:
                 seen.add(interp)
                 yield interp
@@ -173,7 +178,7 @@ def _select_path_interpreter(
 
     # TODO: Allow the selection strategy to be parameterized:
     #   https://github.com/pantsbuild/pex/issues/430
-    return min(candidate_interpreters)
+    return PythonInterpreter.latest_release_of_min_compatible_version(candidate_interpreters)
 
 
 def maybe_reexec_pex(compatibility_constraints=None):