+++ pex/resolver.py
@@ -23,6 +23,8 @@ from pex.pip import PackageIndexConfiguration, get_pip
 from pex.platforms import Platform
 from pex.requirements import local_project_from_requirement, local_projects_from_requirement_file
 from pex.third_party.packaging.markers import Marker
+from pex.third_party.packaging.version import Version
+from pex.third_party.packaging.version import parse as parse_version
 from pex.third_party.pkg_resources import Distribution, Environment, Requirement
 from pex.tracer import TRACER
 from pex.util import CacheHelper
@@ -109,6 +111,19 @@ class DistributionRequirements(object):
     def to_requirement(self, dist):
         req = dist.as_requirement()
 
+        # pkg_resources.Distribution.as_requirement returns requirements in one of two forms:
+        # 1.) project_name==version
+        # 2.) project_name===version
+        # The latter form is used whenever the distribution's version is non-standard. In those
+        # cases we cannot append environment markers since `===` indicates a raw version string to
+        # the right that should not be parsed and instead should be compared literally in full.
+        # See:
+        # + https://www.python.org/dev/peps/pep-0440/#arbitrary-equality
+        # + https://github.com/pantsbuild/pex/issues/940
+        operator, _ = req.specs[0]
+        if operator == "===":
+            return req
+
         markers = OrderedSet()
 
         # Here we map any wheel python requirement to the equivalent environment marker:
@@ -117,11 +132,22 @@ class DistributionRequirements(object):
         # + https://www.python.org/dev/peps/pep-0508/#environment-markers
         python_requires = dist_metadata.requires_python(dist)
         if python_requires:
+
+            def choose_marker(version):
+                # type: (str) -> str
+                parsed_version = parse_version(version)
+                if type(parsed_version) != Version or len(parsed_version.release) > 2:
+                    return "python_full_version"
+                else:
+                    return "python_version"
+
             markers.update(
                 Marker(python_version)
                 for python_version in sorted(
-                    "python_version {operator} {version!r}".format(
-                        operator=specifier.operator, version=specifier.version
+                    "{marker} {operator} {version!r}".format(
+                        marker=choose_marker(specifier.version),
+                        operator=specifier.operator,
+                        version=specifier.version,
                     )
                     for specifier in python_requires
                 )